refactor: Remove unused `SpindexerPositionUtility` instance
This commit removes the unused instance of `SpindexerPositionUtility` from `MainTeleOp`. This utility's methods are static and do not require an object instance for use.

### Key Changes:

*   **`MainTeleOp.java`:**
    *   The `spu` class member has been removed, as its functionality is accessed statically.

*   **`Tune_Spindexer.java`:**
    *   The call to `SpindexerPositionUtility.getNextShootPosition` has been updated to cast the `currentDegrees` variable to an `int`, aligning with the method's expected parameter type.

*   **`HardwareInitializer.java`:**
    *   The redundant `Limelight.initialize(hardwareMap)` call has been removed. The Limelight subsystem is already instantiated directly in `MainTeleOp`.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/OpModes/Tuning/Tune_Spindexer.java
@@ -117,7 +117,7 @@ public class Tune_Spindexer extends OpMode {
 			double currentDegrees = spindexer.getCalibratedPosition();
 
 			// Calculate next shoot position using utility (now handles doubles)
-			double nextShootPosition = SpindexerPositionUtility.getNextShootPosition(currentDegrees);
+			double nextShootPosition = SpindexerPositionUtility.getNextShootPosition((int) currentDegrees);
 
 			// Convert to revolutions for the spindexer (0-1 range)
 			double targetRevolutions = nextShootPosition / 360.0;
